move functions outside of app.vue
