+++ frontend/src/App.vue
@@ -15,78 +15,29 @@ import { onMounted, ref } from 'vue'
 import { setDefaultAuthHeader } from './axiosConfig';
 import { useRouter } from 'vue-router';
 import { getCookie } from './components/cookie_utils';
-import { socket } from '@/socket';
 import Toast from 'primevue/toast'
 import ConfirmDialog from 'primevue/confirmdialog';
 import Menubar from './components/Menubar/Menubar.vue';
 import FriendsMenubar from './components/Friends/FriendsMenubar/FriendsMenubar.vue';
 import GameInvitation from './components/pong/GameInvitation.vue'
 import ChatNotification from './components/Chat/ChatNotification.vue';
-import axiosInstance from './axiosConfig';
+import { loginPlayer } from './components/Auth/initPlayer';
 
 const router = useRouter();
 const sessionVariablesAreSet = ref(false);
 
-// AUTH
-const checkLoggedIn = async () => {
-	const accesstoken = getCookie('auth');
-	if (accesstoken === undefined) {
-		router.push({ name: 'auth' })
-		console.log('not logged?')
-		return;
-	}
-	setDefaultAuthHeader(accesstoken);
-};
-
 onMounted( async () => {
 	const accesstoken = getCookie('auth');
 	if (accesstoken === undefined) {
-		checkLoggedIn();
+		router.push({ name: 'auth' })
 	}
 	else {
 		setDefaultAuthHeader(accesstoken);
-		// load intraname from cookie, done after intra login auth
-		const responsePlayerId = await axiosInstance.get('/user/id');
-		const responseIntraname = await axiosInstance.get('/user/intraname');
-		const responseUsername = await axiosInstance.get('/user/username');
-		const hasAvatar = await axiosInstance.get('player/hasavatar/' + responsePlayerId.data);
-
-		const playerId = parseInt(responsePlayerId.data);
-
-		// set data in session storage
-		sessionStorage.setItem('playerId', responsePlayerId.data);
-		sessionStorage.setItem('intraUsername', responseIntraname.data);
-		sessionStorage.setItem('username', responseUsername.data);
-		sessionVariablesAreSet.value = true;
-
-		// set default avatar
-		if (!hasAvatar.data) {
-			setDefaultAvatar(responsePlayerId.data);
+		if (!localStorage.getItem('logged')) {
+			sessionVariablesAreSet.value = await loginPlayer();
 		}
-
-		socket.auth = { id: playerId };
-		socket.connect();
 	}
-	})
-
-	const setDefaultAvatar = async (playerId: string) => {
-		const randomDecimal = Math.random() ;
-		const randomNumber = Math.floor(randomDecimal * 7) + 1;
-		const defaultAvatarPath = './default_avatars/avatar_' + randomNumber + '.png';
-
-		// Fetch the default avatar file
-		const defaultAvatarFile = await fetch(defaultAvatarPath);
-		const defaultAvatarBlob = await defaultAvatarFile.blob();
-		const defaultAvatar = new File([defaultAvatarBlob], 'default_avatar.png');
-
-		// Send the default avatar file to the server
-		const formData = new FormData();
-		formData.append('avatar', defaultAvatar);
-
-		console.log("avatar: ", defaultAvatarBlob);
-		console.log("playerid: ", playerId);
-		await axiosInstance.post('player/avatar/upload/' + playerId, formData);
-	};
+})
 
 </script>
