+++ frontend/src/components/Auth/AuthCheck.vue
@@ -10,13 +10,12 @@
 		+ process.env.VUE_APP_FT_CLIENTID + '&redirect_uri=' + 
 		process.env.VUE_APP_FT_REDIRECTURI + '&response_type=code';
 		console.log(url);
-	try {
-		window.location.assign(url);
-	} catch (error) {
-		console.log('Error occurred');
+		try {
+			window.location.assign(url);
+		} catch (error) {
+			console.log('Error occurred');
+		}
 	}
-}
-
 </script>
 
 <style scoped>
